getting waiting list with resident teams attached

+++ SocialCareCaseViewerApi/V1/Boundary/Requests/WaitingListRequest.cs
@@ -9,6 +9,6 @@ namespace SocialCareCaseViewerApi.V1.Boundary.Requests
     public class WaitingListRequest
     {
         [FromQuery(Name = "team_id")]
-        public long TeamId { get; set; }
+        public int TeamId { get; set; }
     }
 }
